BlaubaDALI/requirements.txt
# Core dependencies for DALI communication
pyserial>=3.5

# Time handling (Standard Library, no install needed)
# threading (Standard Library, no install needed)